Ultimando detalles

+++ app/Http/Controllers/PatientController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Patient;
+use App\Models\Patient_hospital_medicine;
 use App\Http\Requests\StorePatientRequest;
+use App\Http\Requests\UpdatePatientRequest;
 use App\Http\Resources\PatientResource;
 
 class PatientController extends Controller
@@ -17,7 +19,21 @@ class PatientController extends Controller
     public function index(Request $request)
     {
         $user = $request->user();
-        return PatientResource::collection(Patient::where('id_hospital', $user->id));
+
+        if($request->buscador){
+            $filtro = $request->buscador;
+
+
+
+            return PatientResource::collection(Patient::where('id_hospital', $user->id)
+                         ->where('fullname','LIKE' , '%'.$filtro.'%')
+                         ->orWhere('dni', 'LIKE', '%'.$filtro.'%')
+                         ->paginate(10));
+
+        }
+             
+        return PatientResource::collection(Patient::where('id_hospital', $user->id)->paginate(10) );
+      
         
     }
 
@@ -75,7 +91,19 @@ class PatientController extends Controller
             return abort(403, 'Unauthorized action');
         }
 
+        // Obtenemos el valor de la clave primaria del modelo Medicine
+        $patient_id = $patient->id;
+
+        // Verificamos si hay registros en la tabla Patient_hospital_medicine con el id_medicine igual al valor de la clave primaria del modelo Medicine
+        $patient_hospital_medicine = Patient_hospital_medicine::where('id_medicine', $patient_id)->get();
+
+        // Si hay registros, lanzamos una excepción y no borramos el registro de Medicine
+        if ($patient_hospital_medicine->count() > 0) {
+            throw new \Exception('Cannot delete Patient record because there are related records.');
+        }
+
+        //No hay registros relacionados, asique eliminamos paciente
         $patient->delete();
-        return response('', 204);
+        return response()->json(['success' => 'Record deleted successfully.', 204]);
     }
 }
